[skip ci] Implement the split for xproto.rs, window.rs, and drawable.rs
