+++ src/display/connection/mod.rs
@@ -10,3 +10,11 @@ mod sync;
 #[cfg(feature = "async")]
 pub use async_connection::*;
 pub use sync::*;
+
+use alloc::vec::Vec;
+
+pub(crate) fn standard_fd_warning(fds: &mut Vec<crate::Fd>) {
+    if !fds.is_empty() {
+        log::warn!("Cannot pass file descriptors into request on non-Unix operating systems.");
+    }
+}
